Sprint 2
When a game is started, the user is offered several options: fight, display stats, save, exit or rest. The rest option allows a player to be rested, but cannot be selected during a fight (only between fights). The fight option allows the user to select the attack method they want to launch an attack and then the defense method they'd lik3 to block their opponents attack. A damage is inflicted on the defending player by finding the difference between the attack method's points and defense method's points and subtracting that from the defending player's health points. It is good to note that some attack or defense methods may become unavailable if a player's attack or defense points fall under a certain threshold. When a player's health points reach zero or below that, the player is declared defeated and their data is restored to default to allow a new fight to be initiated.